requirements.txt
absl-py==2.1.0
albucore==0.0.23
albumentations==2.0.0
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
attrs==24.3.0
beautifulsoup4==4.12.3
blis==1.2.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
dm-tree==0.1.8
dnspython==2.7.0
docstring_parser==0.16
efficientnet-pytorch==0.7.1
email_validator==2.2.0
etils==1.11.0
fastai==2.7.18
fastapi==0.115.6
fastapi-cli==0.0.7
fastcore==1.7.28
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.16.1
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2024.12.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
immutabledict==4.2.1
importlib_resources==6.5.2
itsdangerous==2.2.0
jax==0.5.0
jaxlib==0.5.0
Jinja2==3.1.5
joblib==1.4.2
kagglehub==0.3.6
keras==3.8.0
keras-core==0.1.7
keras-cv==0.9.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.20
ml-dtypes==0.4.1
mpmath==1.3.0
munch==4.0.0
murmurhash==1.0.12
mysql-connector-python==9.0.0
mysqlclient==2.2.7
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pillow==11.1.0
preshed==3.0.9
pretrainedmodels==0.7.4
promise==2.3
protobuf==4.25.5
psutil==6.1.1
psycopg2==2.9.10
py-cpuinfo==9.0.0
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.5
pydantic-extra-types==2.10.1
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
retina-face==0.0.17
rich==13.9.4
rich-toolkit==0.12.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
segmentation_models_pytorch==0.4.0
sentencepiece==0.2.0
shellingham==1.5.4
simple-parsing==0.1.6
simsimd==6.2.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.41.3
stringzilla==3.11.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-datasets==4.9.7
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.16.1
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
thinc==8.3.4
threadpoolctl==3.5.0
timm==1.0.13
toml==0.10.2
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
ultralytics==8.3.59
ultralytics-thop==2.0.13
urllib3==2.3.0
uv==0.5.18
uvicorn==0.34.0
wasabi==1.1.3
watchfiles==1.0.4
weasel==0.4.1
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.1
zipp==3.21.0
